fixed fasta

+++ fasta/info.go
@@ -6,9 +6,9 @@ import (
 	"strings"
 )
 
-func Extract(f *Fasta, start int64, end int64, name string) *Fasta {
+func Extract(f *Fasta, start int, end int, name string) *Fasta {
 	var ans Fasta
-	if start < 0 || end < 0 || start > int64(len(f.Seq)) || end > int64(len(f.Seq)) || start > end {
+	if start < 0 || end < 0 || start > len(f.Seq) || end > len(f.Seq) || start > end {
 		log.Fatalf("Invalid start and end given to fasta.Extract. Start: %d, End: %d, ChromLength: %d.", start, end, len(f.Seq))
 	}
 	ans.Seq = dna.Extract(f.Seq, start, end)
